CSHARP-1834: Make minimum value of MaxStaleness 90 seconds.

+++ src/MongoDB.Driver.Core/Core/Operations/QueryHelper.cs
@@ -81,7 +81,7 @@ namespace MongoDB.Driver.Core.Operations
             {
                 { "mode", modeString },
                 { "tags", tagSets, tagSets != null },
-                { "maxStalenessSeconds", () => readPreference.MaxStaleness.Value.TotalSeconds, readPreference.MaxStaleness.HasValue }
+                { "maxStalenessSeconds", () => (int)readPreference.MaxStaleness.Value.TotalSeconds, readPreference.MaxStaleness.HasValue }
             };
         }
     }
